# case-tracking/case-tracking-service-with-rust/openapi.yaml
openapi: 3.1.0
info:
  title: Case Tracking API (Loco edition)
  version: 0.1.0
  description: >
    JSON API for tracking the physical location of paper case-note folders
    across NHS hospital cabinets. The tracker owns no domain tables; every
    entity is proxied from one of five upstream Main-X-Services. See
    `spec/index.md` for the full specification. This document is the
    machine-readable form of `spec/api-contract.md` + `spec/routes.md`.
servers:
  - url: http://localhost:5150
    description: Local dev (JSON API). In dev the SvelteKit app proxies /api here.
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/auth/request:
    post:
      tags: [auth]
      summary: Request a magic-link sign-in email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Always 200 (no email enumeration). `magic_link` is present only in dev/test.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
                  magic_link: { type: [string, 'null'] }
        '422': { $ref: '#/components/responses/Validation' }
  /api/auth/verify:
    post:
      tags: [auth]
      summary: Exchange a magic token for a session (sets the cts_session cookie).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Session established.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: 'cts_session=<jwt>; HttpOnly; SameSite=Lax; Path=/'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserBody' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/me:
    get:
      tags: [auth]
      summary: The signed-in user.
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserBody' }
          description: OK
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cts_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Validation:
      description: Validation failure — each failing field is keyed.
      content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties: { type: string }
    UserBody:
      type: object
      properties:
        user:
          type: object
          properties:
            email: { type: string }
            name: { type: string }
            role: { type: [string, 'null'] }
...